+++ lib/dlms-cosem/include/dlms/cosem/simple_objects.hpp
@@ -914,6 +914,50 @@ private:
   CosemAccessRights rights_;
 };
 
+class CosemSingleActionScheduleObject : public ICosemObject
+{
+public:
+  static const std::uint8_t MaxSupportedVersion = 0u;
+
+  CosemSingleActionScheduleObject(
+    const CosemLogicalName& logicalName,
+    const CosemByteBuffer& executedScript,
+    const CosemByteBuffer& type,
+    const CosemByteBuffer& executionTime,
+    AttributeAccessMode mutableAccess);
+  CosemSingleActionScheduleObject(
+    const CosemLogicalName& logicalName,
+    const CosemByteBuffer& executedScript,
+    const CosemByteBuffer& type,
+    const CosemByteBuffer& executionTime,
+    AttributeAccessMode mutableAccess,
+    std::uint8_t version);
+
+  CosemObjectDescriptor Descriptor() const;
+  CosemAccessRights AccessRights() const;
+  CosemStatus ReadAttribute(
+    std::uint8_t attributeId,
+    CosemByteBuffer& output) const;
+  CosemStatus WriteAttribute(
+    std::uint8_t attributeId,
+    const CosemByteBuffer& input);
+  CosemStatus InvokeMethod(
+    std::uint8_t methodId,
+    const CosemByteBuffer& input,
+    CosemByteBuffer& output);
+
+  const CosemByteBuffer& ExecutedScript() const;
+  const CosemByteBuffer& Type() const;
+  const CosemByteBuffer& ExecutionTime() const;
+
+private:
+  CosemObjectDescriptor descriptor_;
+  CosemByteBuffer executedScript_;
+  CosemByteBuffer type_;
+  CosemByteBuffer executionTime_;
+  CosemAccessRights rights_;
+};
+
 enum class CosemClockBase
 {
   NotDefined = 0,